+++ packages/stream_chat_localizations/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming Beta
+
+- Added translations for new `locationLabel` label.
+
 ## 10.0.0-beta.3
 
 - Included the changes from version [`9.14.0`](https://pub.dev/packages/stream_chat_localizations/changelog).